+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/helper/ModelMapperConverters.java
@@ -1,4 +1,4 @@
-package com.itachallenge.challenge.helpers;
+package com.itachallenge.challenge.helper;
 
 import com.itachallenge.challenge.documents.Challenge;
 import com.itachallenge.challenge.dtos.ChallengeDto;
