+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/service/ChallengeService.java
@@ -2,7 +2,7 @@ package com.itachallenge.challenge.service;
 
 import com.itachallenge.challenge.documents.Challenge;
 import com.itachallenge.challenge.dtos.ChallengeDto;
-import com.itachallenge.challenge.helpers.ModelMapperConverters;
+import com.itachallenge.challenge.helper.ModelMapperConverters;
 import com.itachallenge.challenge.repository.ChallengeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
