+++ itachallenge-challenge/src/test/java/com/itachallenge/challenge/dtos/LanguageDtoTest.java
@@ -2,7 +2,7 @@ package com.itachallenge.challenge.dtos;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.itachallenge.challenge.helpers.ResourceHelper;
+import com.itachallenge.challenge.helper.ResourceHelper;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
