+++ itachallenge-challenge/src/test/java/com/itachallenge/challenge/helpers/ResourceHelperTest.java
@@ -1,5 +1,6 @@
 package com.itachallenge.challenge.helpers;
 
+import com.itachallenge.challenge.helper.ResourceHelper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
